+++ AppleALC/kern_alc.cpp
@@ -211,14 +211,15 @@ void AlcEnabler::updateDeviceProperties(IORegistryEntry *hdaService, DeviceInfo 
 			hdaService->setProperty("alc-layout-id", &layout, sizeof(layout));
 		} else {
 			uint32_t alcId;
-			if (WIOKit::getOSDataValue(hdaService, "alc-layout-id", alcId)) {
-				DBGLOG("alc", "found normal alc-layout-id %u", alcId);
+			if (info->firmwareVendor == DeviceInfo::FirmwareVendor::Apple &&
+				WIOKit::getOSDataValue(hdaService, "alc-layout-id", alcId)) {
+				DBGLOG("alc", "found apple alc-layout-id %u property", alcId);
 			} else if (info->firmwareVendor != DeviceInfo::FirmwareVendor::Apple) {
 				if (WIOKit::getOSDataValue(hdaService, "layout-id", alcId)) {
-					DBGLOG("alc", "found legacy alc-layout-id (from layout-id) %u", alcId);
+					DBGLOG("alc", "found legacy layout-id %u property", alcId);
 					hdaService->setProperty("alc-layout-id", &alcId, sizeof(alcId));
 				} else {
-					SYSLOG("alc", "error: neither alc-layout-id nor layout-id is found in configuration");
+					SYSLOG("alc", "error: no layout-id property found in configuration");
 				}
 			}
 		}