+++ Changelog.md
@@ -2,6 +2,7 @@ AppleALC Changelog
 ==================
 #### v1.5.0
 - Update ALC283 layout-id 88 by xiaoleGun
+- Fixed accidental reading of `alc-layout-id` on non-Apple firmwares
 
 #### v1.4.9
 - Added IDT92HD73E1X5 layout-id 15 for HP Envy h8 1425eg by MacPeet